import mobility and intervention global data

import mobility google reports, acasp and unesco data on interventions

read_google_global()

read_google_region_list()

read_google_region_country(country_iso = "PE")

read_acaps_governments()

read_unesco_education()

Arguments

country_iso
two letter iso code for each country available in read_google_region_list()

Value

import external data sources on covid19 human behaviour or interventions

Functions

read_google_global: import google mobility reports

read_google_region_list: print list of files inside regional zip file

read_google_region_country: import regional mobility data for one country

read_acaps_governments: import acaps dataset

read_unesco_education: import unesco dataset

Examples

library(covid19viz)
library(tidyverse)

# google mobility reports

# global
read_google_global()

#> Parsed with column specification:
#> cols(
#>   country_region_code = col_character(),
#>   country_region = col_character(),
#>   sub_region_1 = col_character(),
#>   sub_region_2 = col_logical(),
#>   metro_area = col_logical(),
#>   iso_3166_2_code = col_character(),
#>   census_fips_code = col_logical(),
#>   date = col_date(format = ""),
#>   retail_and_recreation_percent_change_from_baseline = col_double(),
#>   grocery_and_pharmacy_percent_change_from_baseline = col_double(),
#>   parks_percent_change_from_baseline = col_double(),
#>   transit_stations_percent_change_from_baseline = col_double(),
#>   workplaces_percent_change_from_baseline = col_double(),
#>   residential_percent_change_from_baseline = col_double()
#> )

#> Warning: 2584995 parsing failures.
#>  row        col           expected                  actual                                                                  file
#> 1846 metro_area 1/0/T/F/TRUE/FALSE Kabul Metropolitan Area 'https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv'
#> 1847 metro_area 1/0/T/F/TRUE/FALSE Kabul Metropolitan Area 'https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv'
#> 1848 metro_area 1/0/T/F/TRUE/FALSE Kabul Metropolitan Area 'https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv'
#> 1849 metro_area 1/0/T/F/TRUE/FALSE Kabul Metropolitan Area 'https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv'
#> 1850 metro_area 1/0/T/F/TRUE/FALSE Kabul Metropolitan Area 'https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv'
#> .... .......... .................. ....................... .....................................................................
#> See problems(...) for more details.

#> # A tibble: 2,441,517 x 14
#>    country_region_~ country_region sub_region_1 sub_region_2 metro_area
#>    <chr>            <chr>          <chr>        <lgl>        <lgl>     
#>  1 AE               United Arab E~ NA           NA           NA        
#>  2 AE               United Arab E~ NA           NA           NA        
#>  3 AE               United Arab E~ NA           NA           NA        
#>  4 AE               United Arab E~ NA           NA           NA        
#>  5 AE               United Arab E~ NA           NA           NA        
#>  6 AE               United Arab E~ NA           NA           NA        
#>  7 AE               United Arab E~ NA           NA           NA        
#>  8 AE               United Arab E~ NA           NA           NA        
#>  9 AE               United Arab E~ NA           NA           NA        
#> 10 AE               United Arab E~ NA           NA           NA        
#> # ... with 2,441,507 more rows, and 9 more variables: iso_3166_2_code <chr>,
#> #   census_fips_code <lgl>, date <date>,
#> #   retail_and_recreation_percent_change_from_baseline <dbl>,
#> #   grocery_and_pharmacy_percent_change_from_baseline <dbl>,
#> #   parks_percent_change_from_baseline <dbl>,
#> #   transit_stations_percent_change_from_baseline <dbl>,
#> #   workplaces_percent_change_from_baseline <dbl>,
#> #   residential_percent_change_from_baseline <dbl>

# regional
# first: select country ISO
read_google_region_list()

#> # A tibble: 135 x 4
#>     country_iso Name                                 Length Date               
#>     <chr>       <chr>                                 <dbl> <dttm>             
#>   1 AE          2020_AE_Region_Mobility_Report.csv   121354 2020-09-13 15:27:00
#>   2 AF          2020_AF_Region_Mobility_Report.csv    20553 2020-09-13 15:27:00
#>   3 AG          2020_AG_Region_Mobility_Report.csv    57484 2020-09-13 15:27:00
#>   4 AO          2020_AO_Region_Mobility_Report.csv    58954 2020-09-13 15:27:00
#>   5 AR          2020_AR_Region_Mobility_Report.csv  6932241 2020-09-13 15:27:00
#>   6 AT          2020_AT_Region_Mobility_Report.csv  1402047 2020-09-13 15:27:00
#>   7 AU          2020_AU_Region_Mobility_Report.csv  4329094 2020-09-13 15:27:00
#>   8 AW          2020_AW_Region_Mobility_Report.csv     9967 2020-09-13 15:27:00
#>   9 BA          2020_BA_Region_Mobility_Report.csv    13096 2020-09-13 15:27:00
#>  10 BB          2020_BB_Region_Mobility_Report.csv    94057 2020-09-13 15:27:00
#>  11 BD          2020_BD_Region_Mobility_Report.csv    10939 2020-09-13 15:27:00
#>  12 BE          2020_BE_Region_Mobility_Report.csv   198973 2020-09-13 15:27:00
#>  13 BF          2020_BF_Region_Mobility_Report.csv    58941 2020-09-13 15:27:00
#>  14 BG          2020_BG_Region_Mobility_Report.csv  1486560 2020-09-13 15:27:00
#>  15 BH          2020_BH_Region_Mobility_Report.csv    10633 2020-09-13 15:27:00
#>  16 BJ          2020_BJ_Region_Mobility_Report.csv    62310 2020-09-13 15:27:00
#>  17 BO          2020_BO_Region_Mobility_Report.csv   755614 2020-09-13 15:27:00
#>  18 BR          2020_BR_Region_Mobility_Report.csv 26762400 2020-09-13 15:27:00
#>  19 BS          2020_BS_Region_Mobility_Report.csv    50432 2020-09-13 15:27:00
#>  20 BW          2020_BW_Region_Mobility_Report.csv    75578 2020-09-13 15:27:00
#>  21 BY          2020_BY_Region_Mobility_Report.csv     9768 2020-09-13 15:27:00
#>  22 BZ          2020_BZ_Region_Mobility_Report.csv    70300 2020-09-13 15:27:00
#>  23 CA          2020_CA_Region_Mobility_Report.csv  3845003 2020-09-13 15:27:00
#>  24 CH          2020_CH_Region_Mobility_Report.csv   342762 2020-09-13 15:27:00
#>  25 CI          2020_CI_Region_Mobility_Report.csv   175904 2020-09-13 15:27:00
#>  26 CL          2020_CL_Region_Mobility_Report.csv  1050835 2020-09-13 15:27:00
#>  27 CM          2020_CM_Region_Mobility_Report.csv   275536 2020-09-13 15:27:00
#>  28 CO          2020_CO_Region_Mobility_Report.csv  4094296 2020-09-13 15:27:00
#>  29 CR          2020_CR_Region_Mobility_Report.csv    11287 2020-09-13 15:27:00
#>  30 CV          2020_CV_Region_Mobility_Report.csv    46702 2020-09-13 15:27:00
#>  31 CZ          2020_CZ_Region_Mobility_Report.csv   209675 2020-09-13 15:27:00
#>  32 DE          2020_DE_Region_Mobility_Report.csv   221647 2020-09-13 15:27:00
#>  33 DK          2020_DK_Region_Mobility_Report.csv  1667194 2020-09-13 15:27:00
#>  34 DO          2020_DO_Region_Mobility_Report.csv   504999 2020-09-13 15:27:00
#>  35 EC          2020_EC_Region_Mobility_Report.csv  1614049 2020-09-13 15:27:00
#>  36 EE          2020_EE_Region_Mobility_Report.csv   160004 2020-09-13 15:27:00
#>  37 EG          2020_EG_Region_Mobility_Report.csv   426886 2020-09-13 15:27:00
#>  38 ES          2020_ES_Region_Mobility_Report.csv   891125 2020-09-13 15:27:00
#>  39 FI          2020_FI_Region_Mobility_Report.csv  1126790 2020-09-13 15:27:00
#>  40 FJ          2020_FJ_Region_Mobility_Report.csv     9406 2020-09-13 15:27:00
#>  41 FR          2020_FR_Region_Mobility_Report.csv  1712923 2020-09-13 15:27:00
#>  42 GA          2020_GA_Region_Mobility_Report.csv    40689 2020-09-13 15:27:00
#>  43 GB          2020_GB_Region_Mobility_Report.csv  7153647 2020-09-13 15:27:00
#>  44 GE          2020_GE_Region_Mobility_Report.csv    19975 2020-09-13 15:27:00
#>  45 GH          2020_GH_Region_Mobility_Report.csv    38477 2020-09-13 15:27:00
#>  46 GR          2020_GR_Region_Mobility_Report.csv   153113 2020-09-13 15:27:00
#>  47 GT          2020_GT_Region_Mobility_Report.csv   345892 2020-09-13 15:27:00
#>  48 GW          2020_GW_Region_Mobility_Report.csv     9507 2020-09-13 15:27:00
#>  49 HK          2020_HK_Region_Mobility_Report.csv    10867 2020-09-13 15:27:00
#>  50 HN          2020_HN_Region_Mobility_Report.csv   785765 2020-09-13 15:27:00
#>  51 HR          2020_HR_Region_Mobility_Report.csv   310498 2020-09-13 15:27:00
#>  52 HT          2020_HT_Region_Mobility_Report.csv   115294 2020-09-13 15:27:00
#>  53 HU          2020_HU_Region_Mobility_Report.csv   273940 2020-09-13 15:27:00
#>  54 ID          2020_ID_Region_Mobility_Report.csv   482720 2020-09-13 15:27:00
#>  55 IE          2020_IE_Region_Mobility_Report.csv   352783 2020-09-13 15:27:00
#>  56 IL          2020_IL_Region_Mobility_Report.csv   273435 2020-09-13 15:27:00
#>  57 IN          2020_IN_Region_Mobility_Report.csv  8726568 2020-09-13 15:27:00
#>  58 IQ          2020_IQ_Region_Mobility_Report.csv     9797 2020-09-13 15:27:00
#>  59 IT          2020_IT_Region_Mobility_Report.csv  2016358 2020-09-13 15:27:00
#>  60 JM          2020_JM_Region_Mobility_Report.csv   193826 2020-09-13 15:27:00
#>  61 JO          2020_JO_Region_Mobility_Report.csv   180433 2020-09-13 15:27:00
#>  62 JP          2020_JP_Region_Mobility_Report.csv   549338 2020-09-13 15:27:00
#>  63 KE          2020_KE_Region_Mobility_Report.csv   494727 2020-09-13 15:27:00
#>  64 KG          2020_KG_Region_Mobility_Report.csv   105688 2020-09-13 15:27:00
#>  65 KH          2020_KH_Region_Mobility_Report.csv   278775 2020-09-13 15:27:00
#>  66 KR          2020_KR_Region_Mobility_Report.csv    25720 2020-09-13 15:27:00
#>  67 KW          2020_KW_Region_Mobility_Report.csv   120968 2020-09-13 15:27:00
#>  68 KZ          2020_KZ_Region_Mobility_Report.csv    10613 2020-09-13 15:27:00
#>  69 LA          2020_LA_Region_Mobility_Report.csv     9282 2020-09-13 15:27:00
#>  70 LB          2020_LB_Region_Mobility_Report.csv   426603 2020-09-13 15:27:00
#>  71 LI          2020_LI_Region_Mobility_Report.csv     9769 2020-09-13 15:27:00
#>  72 LK          2020_LK_Region_Mobility_Report.csv    10631 2020-09-13 15:27:00
#>  73 LT          2020_LT_Region_Mobility_Report.csv   149896 2020-09-13 15:27:00
#>  74 LU          2020_LU_Region_Mobility_Report.csv    10985 2020-09-13 15:27:00
#>  75 LV          2020_LV_Region_Mobility_Report.csv   410968 2020-09-13 15:27:00
#>  76 LY          2020_LY_Region_Mobility_Report.csv   113797 2020-09-13 15:27:00
#>  77 MA          2020_MA_Region_Mobility_Report.csv    56703 2020-09-13 15:27:00
#>  78 MD          2020_MD_Region_Mobility_Report.csv    10121 2020-09-13 15:27:00
#>  79 MK          2020_MK_Region_Mobility_Report.csv    11810 2020-09-13 15:27:00
#>  80 ML          2020_ML_Region_Mobility_Report.csv    23432 2020-09-13 15:27:00
#>  81 MM          2020_MM_Region_Mobility_Report.csv    45716 2020-09-13 15:27:00
#>  82 MN          2020_MN_Region_Mobility_Report.csv    23672 2020-09-13 15:27:00
#>  83 MT          2020_MT_Region_Mobility_Report.csv     9733 2020-09-13 15:27:00
#>  84 MU          2020_MU_Region_Mobility_Report.csv   153612 2020-09-13 15:27:00
#>  85 MX          2020_MX_Region_Mobility_Report.csv   430010 2020-09-13 15:27:00
#>  86 MY          2020_MY_Region_Mobility_Report.csv   222230 2020-09-13 15:27:00
#>  87 MZ          2020_MZ_Region_Mobility_Report.csv   283832 2020-09-13 15:27:00
#>  88 NA          2020_NA_Region_Mobility_Report.csv    92487 2020-09-13 15:27:00
#>  89 NE          2020_NE_Region_Mobility_Report.csv    72034 2020-09-13 15:27:00
#>  90 NG          2020_NG_Region_Mobility_Report.csv  2182993 2020-09-13 15:27:00
#>  91 NI          2020_NI_Region_Mobility_Report.csv   220886 2020-09-13 15:27:00
#>  92 NL          2020_NL_Region_Mobility_Report.csv  4813185 2020-09-13 15:27:00
#>  93 NO          2020_NO_Region_Mobility_Report.csv  1732666 2020-09-13 15:27:00
#>  94 NP          2020_NP_Region_Mobility_Report.csv    10088 2020-09-13 15:27:00
#>  95 NZ          2020_NZ_Region_Mobility_Report.csv   227132 2020-09-13 15:27:00
#>  96 OM          2020_OM_Region_Mobility_Report.csv   171747 2020-09-13 15:27:00
#>  97 PA          2020_PA_Region_Mobility_Report.csv    26650 2020-09-13 15:27:00
#>  98 PE          2020_PE_Region_Mobility_Report.csv  1734899 2020-09-13 15:27:00
#>  99 PG          2020_PG_Region_Mobility_Report.csv    11642 2020-09-13 15:27:00
#> 100 PH          2020_PH_Region_Mobility_Report.csv   404516 2020-09-13 15:27:00
#> 101 PK          2020_PK_Region_Mobility_Report.csv   309845 2020-09-13 15:27:00
#> 102 PL          2020_PL_Region_Mobility_Report.csv  6269809 2020-09-13 15:27:00
#> 103 PR          2020_PR_Region_Mobility_Report.csv   734904 2020-09-13 15:27:00
#> 104 PT          2020_PT_Region_Mobility_Report.csv  2502262 2020-09-13 15:27:00
#> 105 PY          2020_PY_Region_Mobility_Report.csv   261385 2020-09-13 15:27:00
#> 106 QA          2020_QA_Region_Mobility_Report.csv    23947 2020-09-13 15:27:00
#> 107 RE          2020_RE_Region_Mobility_Report.csv    26381 2020-09-13 15:27:00
#> 108 RO          2020_RO_Region_Mobility_Report.csv  3137494 2020-09-13 15:27:00
#> 109 RS          2020_RS_Region_Mobility_Report.csv    19783 2020-09-13 15:27:00
#> 110 RU          2020_RU_Region_Mobility_Report.csv   250859 2020-09-13 15:27:00
#> 111 RW          2020_RW_Region_Mobility_Report.csv   138669 2020-09-13 15:27:00
#> 112 SA          2020_SA_Region_Mobility_Report.csv   202298 2020-09-13 15:27:00
#> 113 SE          2020_SE_Region_Mobility_Report.csv  3639384 2020-09-13 15:27:00
#> 114 SG          2020_SG_Region_Mobility_Report.csv    10988 2020-09-13 15:27:00
#> 115 SI          2020_SI_Region_Mobility_Report.csv  1526667 2020-09-13 15:27:00
#> 116 SK          2020_SK_Region_Mobility_Report.csv  1239019 2020-09-13 15:27:00
#> 117 SN          2020_SN_Region_Mobility_Report.csv   150596 2020-09-13 15:27:00
#> 118 SV          2020_SV_Region_Mobility_Report.csv   229173 2020-09-13 15:27:00
#> 119 TG          2020_TG_Region_Mobility_Report.csv    37759 2020-09-13 15:27:00
#> 120 TH          2020_TH_Region_Mobility_Report.csv    10400 2020-09-13 15:27:00
#> 121 TJ          2020_TJ_Region_Mobility_Report.csv    26758 2020-09-13 15:27:00
#> 122 TR          2020_TR_Region_Mobility_Report.csv  6238280 2020-09-13 15:27:00
#> 123 TT          2020_TT_Region_Mobility_Report.csv    12879 2020-09-13 15:27:00
#> 124 TW          2020_TW_Region_Mobility_Report.csv     9450 2020-09-13 15:27:00
#> 125 TZ          2020_TZ_Region_Mobility_Report.csv   651658 2020-09-13 15:27:00
#> 126 UA          2020_UA_Region_Mobility_Report.csv    24207 2020-09-13 15:27:00
#> 127 UG          2020_UG_Region_Mobility_Report.csv   241016 2020-09-13 15:27:00
#> 128 US          2020_US_Region_Mobility_Report.csv 40702438 2020-09-13 15:27:00
#> 129 UY          2020_UY_Region_Mobility_Report.csv   304201 2020-09-13 15:27:00
#> 130 VE          2020_VE_Region_Mobility_Report.csv    10968 2020-09-13 15:27:00
#> 131 VN          2020_VN_Region_Mobility_Report.csv   839784 2020-09-13 15:27:00
#> 132 YE          2020_YE_Region_Mobility_Report.csv     9148 2020-09-13 15:27:00
#> 133 ZA          2020_ZA_Region_Mobility_Report.csv   140269 2020-09-13 15:27:00
#> 134 ZM          2020_ZM_Region_Mobility_Report.csv    23707 2020-09-13 15:27:00
#> 135 ZW          2020_ZW_Region_Mobility_Report.csv   254493 2020-09-13 15:27:00

# second: read specific country data
peru <- read_google_region_country(country_iso = "PE")

#> Parsed with column specification:
#> cols(
#>   country_region_code = col_character(),
#>   country_region = col_character(),
#>   sub_region_1 = col_character(),
#>   sub_region_2 = col_character(),
#>   metro_area = col_logical(),
#>   iso_3166_2_code = col_character(),
#>   census_fips_code = col_logical(),
#>   date = col_date(format = ""),
#>   retail_and_recreation_percent_change_from_baseline = col_double(),
#>   grocery_and_pharmacy_percent_change_from_baseline = col_double(),
#>   parks_percent_change_from_baseline = col_double(),
#>   transit_stations_percent_change_from_baseline = col_double(),
#>   workplaces_percent_change_from_baseline = col_double(),
#>   residential_percent_change_from_baseline = col_double()
#> )

peru %>% count(sub_region_1)

#> # A tibble: 27 x 2
#>    sub_region_1      n
#>    <chr>         <int>
#>  1 Amazonas       1050
#>  2 Ancash         1871
#>  3 Apurimac        816
#>  4 Arequipa       1470
#>  5 Ayacucho       1050
#>  6 Cajamarca      1850
#>  7 Callao Region   210
#>  8 Cusco          1668
#>  9 Huancavelica    630
#> 10 Huanuco         630
#> # ... with 17 more rows

# acaps data
read_acaps_governments()

#> # A tibble: 193 x 8
#>    Country_name ISO3  REGION Measure_taxonomy Measure_Category SOURCE_TYPE
#>    <chr>        <chr> <chr>  <chr>            <chr>            <chr>      
#>  1 Afghanistan  AFG   Asia   Schools closure  Social distanci~ Government 
#>  2 Albania      ALB   Europe Border checks    Movement restri~ Media      
#>  3 Algeria      DZA   Africa Border closure   Movement restri~ UN         
#>  4 Angola       AGO   Africa Closure of busi~ Social distanci~ Other orga~
#>  5 Antigua and~ ATG   Ameri~ Isolation and q~ Public health m~ Social med~
#>  6 Argentina    ARG   Ameri~ Limit public ga~ Social distanci~ Other      
#>  7 Armenia      ARM   Asia   Economic measur~ Governance and ~ NA         
#>  8 Australia    AUS   Pacif~ Awareness campa~ Public health m~ NA         
#>  9 Austria      AUT   Europe International f~ Movement restri~ NA         
#> 10 Azerbaijan   AZE   Asia   Health screenin~ Public health m~ NA         
#> # ... with 183 more rows, and 2 more variables: Enforcement_taxonomy <chr>,
#> #   Log_type <chr>

# unesco education data
read_unesco_education()

#> Parsed with column specification:
#> cols(
#>   Date = col_character(),
#>   ISO = col_character(),
#>   Country = col_character(),
#>   Status = col_character(),
#>   Note = col_logical()
#> )

#> # A tibble: 41,714 x 5
#>    Date       ISO   Country  Status                 Note 
#>    <chr>      <chr> <chr>    <chr>                  <lgl>
#>  1 17/02/2020 CHN   China    Partially open         NA   
#>  2 17/02/2020 MNG   Mongolia Closed due to COVID-19 NA   
#>  3 18/02/2020 CHN   China    Partially open         NA   
#>  4 18/02/2020 MNG   Mongolia Closed due to COVID-19 NA   
#>  5 19/02/2020 CHN   China    Partially open         NA   
#>  6 19/02/2020 MNG   Mongolia Closed due to COVID-19 NA   
#>  7 20/02/2020 CHN   China    Partially open         NA   
#>  8 20/02/2020 MNG   Mongolia Closed due to COVID-19 NA   
#>  9 21/02/2020 CHN   China    Closed due to COVID-19 NA   
#> 10 21/02/2020 MNG   Mongolia Closed due to COVID-19 NA   
#> # ... with 41,704 more rows